requirements.txt
cffi==1.15.1
cli-exit-tools==1.2.3.2
click==8.1.3
config==0.5.1
cryptography==39.0.0
fake-winreg==1.6.2.2
lib-detect-testenv==2.0.3
lib-platform==1.2.7
lib-registry==2.0.7
oracledb==1.2.2
pycparser==2.21
python-dotenv==0.21.1
wrapt==1.14.1
camelot-py==0.9.0
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
cli-exit-tools==1.2.3.2
click==8.1.3
config==0.5.1
cryptography==39.0.0
distro==1.8.0
et-xmlfile==1.1.0
fake-winreg==1.6.2.2
lib-detect-testenv==2.0.3
lib-platform==1.2.7
lib-registry==2.0.7
numpy==1.24.2
opencv-python==4.7.0.72
openpyxl==3.1.2
oracledb==1.2.2
pandas==1.5.3
pdfminer.six==20221105
pycparser==2.21
pypdf2==3.0.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.2
PyYAML==6.0
six==1.16.0
tabula==1.0.5
tabula-py==2.7.0
typing-extensions==4.5.0
wrapt==1.14.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cli-exit-tools==1.2.3.2
click==8.1.3
colorama==0.4.6
config==0.5.1
cryptography==39.0.0
distro==1.8.0
fake-winreg==1.6.2.2
idna==3.4
lib-detect-testenv==2.0.3
lib-platform==1.2.7
lib-registry==2.0.7
numpy==1.24.2
oracledb==1.2.2
pandas==1.5.3
pycparser==2.21
pypdf2==3.0.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3
requests==2.28.2
scipy==1.10.1
six==1.16.0
soxr==0.3.4
SpeechRecognition==3.10.0
tabula==1.0.5
tabula-py==2.7.0
typing-extensions==4.5.0
urllib3==1.26.15
wrapt==1.14.1
